requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
